Move `types.ts` to vendor sub-directory.
